fix: defer foreground refresh while hosted auth is pending
Returning from the Chrome Custom Tab foregrounds the app, which triggers
ConfigurationManager's lifecycle callback. That callback defers its refresh
while a browser-based auth flow is still in flight, but the guard only
consulted SSOService. Hosted auth was omitted, so the refresh ran unguarded,
rotated the device token, and the in-flight redemption then failed its
response-freshness check in HostedAuthService.redeemAndComplete.

Users saw the sign-in fail with "Hosted auth redemption response is no longer
current." It is a race, so it only reproduced when enough time passed in the
browser for the foreground refresh to win — reliably reproducible by waiting
~45s on the Account Portal OTP screen before submitting the code.

Add HostedAuthService.hasPendingAuthentication() to the guard, and rename
hasPendingSsoFlow/deferForegroundRefreshDuringPendingSso to
hasPendingAuthFlow/deferForegroundRefreshDuringPendingAuth since they now
cover more than SSO. Both are internal/private, so the rename is contained.
